Refactor producer and highlevelproducer to share the same code
